Initial version of the Grails FileDialog prototype, which includes the ability to browse server side files.
At the current release, the dialog itself is not properly modal, and is simply pasted on top of the page.

Also, it incorrectly displays file options that cannot be browsed due to permissions.  While files and folders are still inaccessible, their names are displayed.

Finally, the dialog does not yet allow the selection of folders for destinations.  This should be changed shortly.